Automated Calendar.
This is a program that creates the folders for each fiscal year of a given company. Inside we will find 12 other folders named after each month of the year, this works for keeping track of the activity in our company in a quick and organized way by redirecting the documents.